added controller for routing
