+++ client/package.json
@@ -7,6 +7,7 @@
     "axios": "^0.18.0",
     "react": "^16.6.1",
     "react-dom": "^16.6.1",
+    "react-router-dom": "^5.1.2",
     "react-scripts": "2.1.1"
   },
   "scripts": {
